Refactor IExtendedBlockStorageMixin
